+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 *  Fix a bug in CF Table that when creating a safety constraint for every hazard of an uca the hazard title is changed to no hazard
 *  Fix a bug in CF Table if I link a SC to one from previous step and afterwards I want to delete it, the operation is not performed.
 *  Fix a bug in CF Table newly added SC aren't linked with any CF in CausalAnalysis:Safety Constraints Step 4
+*  Fix a bug in the collaboration system that the system description is not synchronized
+*  Fix a bug in the collaboration system that no links for the corresponding safety constraints/ design requirements are imported
 
 ### Other Changes
 *  Adding of new Control Actions is now only possible in the Control Structure to prevent inconsistencies between step 2 and 3